MLOpsminiproject/requirements.txt
torch
torchvision
numpy
pandas
mlflow
mediapipe==0.10.14
Pillow
boto3
opencv-python
